+++ modules/talos-cluster/main.tf
@@ -1,7 +1,7 @@
 terraform {
   required_providers {
     talos = {
-      source  = "siderolabs/talos"
+      source = "siderolabs/talos"
       # TODO: using alpha to fix "inconsistent final plan" bug (siderolabs/terraform-provider-talos#352).
       # Revert to stable when v0.12.0 is released.
       version = "0.12.0-alpha.5"